erp_conexion/requirements.txt
pyodbc
polars[excel]
python-dotenv
